+++ lib/rails_admin_settings/settings.rb
@@ -66,7 +66,11 @@ class Settings < BasicObject
     def apply_defaults!(file, verbose = false)
       if File.file?(file)
         puts "[settings] Loading from #{file}" if verbose
-        yaml = YAML.load(File.read(file), safe: true)
+        if defined?(Psych) && Psych.respond_to?(:safe_load)
+          yaml = Psych.safe_load(File.read(file))
+        else
+          yaml = YAML.load(File.read(file), safe: true)
+        end
         yaml.each_pair do |namespace, vals|
           vals.symbolize_keys!
           n = ns(namespace)