+++ spec/defaults_spec.rb
@@ -52,12 +52,12 @@ describe 'Settings loading defaults' do
     expect(Settings.ns(:main).footer).to eq 'test <b></b>'
     expect(Settings.footer).to eq 'zzz'
 
-    Settings.ns(:main).phone.should eq '906 1111111'
-    Settings.ns(:main).true_setting.should be true
-    Settings.ns(:main).false_setting.should be false
-    Settings.ns(:other).footer.should eq 'zzz'
-    Settings.ns(:main).footer.should eq 'test <b></b>'
-    Settings.footer.should eq 'zzz'
+    expect(Settings.ns(:main).phone).to eq '906 1111111'
+    expect(Settings.ns(:main).true_setting).to be true
+    expect(Settings.ns(:main).false_setting).to be false
+    expect(Settings.ns(:other).footer).to eq 'zzz'
+    expect(Settings.ns(:main).footer).to eq 'test <b></b>'
+    expect(Settings.footer).to eq 'zzz'
 
     expect(Settings.ns(:etc, fallback: :main).phone).to eq '906 1111111'
     expect(Settings.ns(:etc, fallback: :main).footer).to eq 'test <b></b>'