+++ spec/factories/setting.rb
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :setting, class: RailsAdminSettings::Setting do
     sequence(:key){|n| "setting_#{n}" }
     raw "Контент 1"