+++ spec/migration_spec.rb
@@ -11,10 +11,10 @@ describe 'Migrating from old versions' do
       coll.insert({enabled: true, key: 'test', raw: '9060000000', type: 'phone'})
     end
     RailsAdminSettings.migrate!
-    RailsAdminSettings::Setting.first.key.should eq 'test'
-    RailsAdminSettings::Setting.first.raw.should eq '9060000000'
-    RailsAdminSettings::Setting.first.ns.should eq 'main'
-    RailsAdminSettings::Setting.first.kind.should eq 'phone'
+    expect(RailsAdminSettings::Setting.first.key).to eq 'test'
+    expect(RailsAdminSettings::Setting.first.raw).to eq '9060000000'
+    expect(RailsAdminSettings::Setting.first.ns).to eq 'main'
+    expect(RailsAdminSettings::Setting.first.kind).to eq 'phone'
   end
 end